更新 课程列表节课程收听人数统计

+++ model/src/main/java/com/zwl/model/po/ClassSetStatistics.java
@@ -9,7 +9,7 @@ public class ClassSetStatistics {
 
     private Long classSetId;
 
-    private Integer browseCount;
+    private Long browseCount;
 
     private Integer shareCount;
